requirements.txt
deepeval
google-genai
dotenv